pyproject.toml
[tool.poetry]
name = "ls-converter"
version = "0.0.1"
description = "Creates ready-to-use Label Studio pre-populated JSON files from popular OCR formats."
authors = ["Kalle Westerling <kalle.westerling@bl.uk>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Living-with-machines/label-studio-converter"
repository = "https://github.com/Living-with-machines/label-studio-converter"
keywords = ["computer-vision", "deep-learning", "image-annotation", "annotation", "annotations", "dataset", "image-classification", "labeling", "datasets", "semantic-segmentation", "annotation-tool", "text-annotation", "boundingbox", "image-labeling", "labeling-tool", "image-labelling-tool", "data-labeling", "label-studio", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Image Processing",
]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^9.3.0"
requests = "^2.28.1"
xmltodict = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.12.0"
xmltodict = "^0.13.0"
pytesseract = "^0.3.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
